identifiserte hva flagg er og omplasserte den, omnavnet den og fjernet litt graps.
